minor fixes: - rename for consistency - rearrange SceneNodeDataImpl fields - typo
