+++ scene/core/src/light.rs
